refactor: `integration/` directory cleanup
